Use JMH to benchmark for Scala implementation.
Previously, the Scala implementation was benchmarked with
`System.nanoTime`, which is known to produce unreliable numbers on the
JVM. This commit configures the build to benchmark the Scala
implementation with JMH instead. JMH is specifically designed to run
microbenchmarks on the JVM taking into account warmup time and other
factors.

The command `sbt bench` runs three warmup iterations, three benchmark
iterations and finally print out a report like this:
```
Benchmark        (height)  (scene)  (width)  Mode  Cnt     Score      Error  Units
Bench.construct      1000   rgbbox     1000  avgt    3     0.246 ±    0.134  ms/op
Bench.construct      1000    irreg     1000  avgt    3     8.184 ±    1.626  ms/op
Bench.render         1000   rgbbox     1000  avgt    3  2862.550 ± 2098.730  ms/op
Bench.render         1000    irreg     1000  avgt    3  1327.580 ±  671.079  ms/op
```

The performance is much better when using GraalVM 20.0.0 instead of
AdoptOpenJDK 8.
```
❯ jabba use graalvm@20.0.0
❯ sbt bench
...
Benchmark        (scene)  Mode  Cnt    Score     Error  Units
Bench.construct   rgbbox  avgt    3    0.229 ±   0.014  ms/op
Bench.construct    irreg  avgt    3    6.187 ±   0.540  ms/op
Bench.render      rgbbox  avgt    3  824.626 ± 449.169  ms/op
Bench.render       irreg  avgt    3  277.417 ±  39.216  ms/op
```

I tried running the non-JMH benchmarks with GraalVM native-image but
got less promising results.
```
❯ jabba use graalvm@20.0.0
❯ gu install native-image
❯ sbt publishLocal
❯ native-image \
    -cp $(cs fetch --classpath raytracer:raytracer_2.13:0.1.0-SNAPSHOT) \
    --initialize-at-build-time \
    raytracer.Main \
    raytracer
❯ ./raytracer
Error parsing command line arguments, running rgbbox 1000x1000 px
Constructing the scene took: 1.827164ms (average over 100 repetitions)
Rendering took: 2814.0152406ms (average over 5 repetitions)
wrote output to 'out.ppm'
```

Tools used:

* `jabba`: https://github.com/shyiko/jabba
* `cs`: https://get-coursier.io/
