+++ scala/src/main/scala/raytracer/Raytracer.scala
@@ -1,8 +1,6 @@
 package raytracer
 
-import scala.util.Try
-
-object Raytracer extends App {
+object Raytracer {
 
   type Pos = Vec3
   type Dir = Vec3
@@ -157,28 +155,4 @@ object Raytracer extends App {
     // First result
     timingHack.head._2
   }
-
-  val (scene, width, height) = (args match {
-    case Array(s, w, h, _*) => for {
-      ss <- Scene.fromString(s)
-      ww <- Try(w.toInt).toOption
-      hh <- Try(h.toInt).toOption
-    } yield (ss, ww, hh)
-    case _ => None
-  }).getOrElse({
-    println("Error parsing command line arguments, running rgbbox 1000x1000 px")
-    (Scene.rgbbox, 1000, 1000)
-  })
-
-  import scala.concurrent.ExecutionContext.Implicits.global
-
-  val (objs, cam) = time("Constructing the scene took", 100, scene.toObjsCam(width, height))
-  val out = time("Rendering took", 5, render(objs, width, height, cam))
-
-  import java.io._
-  val pw = new PrintWriter(new File("out.ppm"))
-  pw.write(out.toPPM)
-  println("wrote output to 'out.ppm'")
-  pw.close
-
 }
